+++ packages/r/openproblems/R/find_project_root.R
@@ -23,5 +23,5 @@ find_project_root <- function(path) {
     return(NULL)
   }
 
-  path
+  as.character(path)
 }